requirements.txt
Django>=5.0,<5.1                        # En güncel Django LTS veya stabil sürümü
psycopg2-binary>=2.9.9,<3.0             # psycopg2'nin güncel ve binary paketi
python-dotenv>=1.0.0
uWSGI>=2.0.23,<2.1                      # En güncel stabil uWSGI sürümü
djangorestframework>=3.14,<3.15         # Django REST Framework (API için gerekli)
pillow==10.2.0                          # Görüntü işleme kütüphanesi (django-imagekit için bağımlılık)
django-imagekit==5.0                    # Django için görüntü işleme ve dönüştürme kütüphanesi
pillow-heif                             # AVIF desteği
django-cleanup                            
django-tinymce                          # Text Field için
django-allauth==0.61.1